% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocher.R
\name{tocher}
\alias{tocher}
\title{Tocher Clustering}
\usage{
tocher(d, algorithm = c("original", "sequential"))
}
\arguments{
\item{d}{distance matrix, a dist "object"}

\item{algorithm}{either "percentual" (default) or "proportion"}
}
\description{
This function computes clusters based on a distance matrix
}
\examples{
\dontrun{
tocher(d = dist(rnorm(10)))
tocher(algorithm = "sequential")
}
}
\author{
Anderson R da Silva
}
